refactor: fix -Wconversion sign-conversion warnings
- formatter.c: cast ret to size_t before subtraction with remaining (size_t)
- dispatcher.c: cast sink_count to size_t in realloc size expressions
- dispatcher.c: use size_t for ansi_count comparison to avoid enum sign conversion
- async.c: cast queue_size to size_t when passed to mpsc_queue_create
- socket_sink.c: cast port to in_port_t for htons; use ssize_t for send() return

+++ core/async.c
@@ -98,7 +98,7 @@ int log_async_init(int queue_size) {
     if (g_async_logger.running)
         return 0;
 
-    g_async_logger.queue = mpsc_queue_create(queue_size);
+    g_async_logger.queue = mpsc_queue_create((size_t)queue_size);
     if (!g_async_logger.queue)
         return -1;
